Refactor: Plugin Part 5
