+++ internal/common/process_step.go
@@ -3,9 +3,9 @@ package common
 import (
 	"time"
 
-	"github.com/AlertFlow/runner/internal/plugin"
 	"github.com/AlertFlow/runner/pkg/executions"
 	"github.com/AlertFlow/runner/pkg/models"
+	"github.com/AlertFlow/runner/pkg/plugin"
 
 	log "github.com/sirupsen/logrus"
 )
