+++ pkg/plugin/core.go
