+++ pkg/plugin/init.go
