+++ pkg/plugin/manager.go
