+++ pkg/plugin/manifest.go
